requirements.txt
# Sistema Multi-Agente TechCorp — Dependencias
# Instalar con: pip install -r requirements.txt

# ── LLM y Embeddings ──────────────────────────────────────────────────────────
openai>=1.55.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
tiktoken>=0.8.0

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Observabilidad ────────────────────────────────────────────────────────────
langfuse>=2.50.0

# ── Entorno y Configuración ───────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Jupyter ───────────────────────────────────────────────────────────────────
jupyter>=1.1.0
ipykernel>=6.29.0
ipywidgets>=8.1.0

# ── Utilitarios ───────────────────────────────────────────────────────────────
tqdm>=4.66.0
pandas>=2.2.0
